requirements.txt
apache-airflow==2.7.0
requests
pandas
sqlalchemy
psycopg2-binary
clickhouse-driver
